Better TS type formatting for array types
